misc: added generic AttributedEnumInspector
