+++ Editor/Inspectors/AttributedEnumInspector.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c60df3d32f18d4d3e9019e2592b52ba9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
