+++ Editor/Inspectors/LampDataTypeInspector.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c3c891d4e51ab4d26aec2424438cc7d8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
